% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.sql}
\alias{is.sql}
\title{Check whether an object is a \code{sql} query}
\usage{
is.sql(x)
}
\arguments{
\item{An}{R object.}
}
\value{
A \code{logical(1)} vector.
}
\description{
Check whether an object is a \code{sql} query
}
