% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.tbl_spark}
\alias{is.tbl_spark}
\title{Check whether an object is a \code{tbl_spark} object}
\usage{
is.tbl_spark(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
A \code{logical(1)} vector.
}
\description{
Check whether an object is a \code{tbl_spark} object
}
