% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_open_spark_connection}
\alias{is_open_spark_connection}
\title{Check A Spark Object's Validity}
\usage{
is_open_spark_connection(sc)
}
\arguments{
\item{sc}{An R object}
}
\value{
\code{logical(1)}
}
\description{
Check whether an R object is a valid, open Spark connection
}
