% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_spark_tbl}
\alias{is_spark_tbl}
\title{Is \code{spark_tbl}}
\usage{
is_spark_tbl(what)
}
\arguments{
\item{what}{An R object}
}
\value{
\code{logical(1)}
}
\description{
Check if an R object is a reference to a valid Spark DataFrame
}
