% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_from_spark}
\alias{read_from_spark}
\title{Read from Spark}
\usage{
read_from_spark(what, sc, ...)
}
\arguments{
\item{what}{Either a \code{tbl_spark} or a SQL query \code{character(1)} string.}

\item{sc}{A \code{spark_connection} used in conjunction with a SQL query given to \code{what}.}

\item{...}{Additional parameters passed on to \code{\link[dplyr:collect]{dplyr::collect()}} or \code{\link[DBI:dbGetQuery]{DBI::dbGetQuery()}}}
}
\value{
A tibble.
}
\description{
Read data from a Spark connection into R either by reference or by query.
}
\examples{
sc <- sparklyr::spark_connect(master = "local")
mt_spark <- write_to_spark(what = mtcars, sc = sc, name = "mtcars")
read_from_spark(what = mt_spark)
read_from_spark(what = "SELECT * FROM mtcars", sc = sc)

}
\seealso{
\code{\link[dplyr:collect]{dplyr::collect()}}, \code{\link[DBI:dbGetQuery]{DBI::dbGetQuery()}}
}
