% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_collect_data.R
\name{spark_collect_data}
\alias{spark_collect_data}
\title{Collect a Spark DataFrame into R}
\usage{
spark_collect_data(x, sc = NULL)
}
\arguments{
\item{x}{One of:
\itemize{
\item A \code{tbl_spark} object.
\item A \code{character(1)} name of the table to retrieve passed with \code{sc}.
}}

\item{sc}{A \code{spark_connection}, only required if \code{x} is passed as a
\code{character(1)} (def: \code{NULL}).}
}
\value{
A \code{tibble}.
}
\description{
Given the name of a Spark DataFrame or a reference to a Spark DataFrame,
collect it into R from a \code{spark_connection}.
}
\examples{
\dontrun{
sc <- sparklyr::spark_connect(master = "local")
mtcars_spark <- dplyr::copy_to(sc, mtcars)

# We can collect the data using the R objet
spark_collect_data(mtcars_spark)

# Or by referencing the name within the spark connection
spark_collect_data(x = "mtcars", sc = sc)
}

}
