% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_pull_data.R
\name{spark_pull_data}
\alias{spark_pull_data}
\title{Pull a Spark DataFrame into R}
\usage{
spark_pull_data(x, sc = NULL)
}
\arguments{
\item{x}{Either the name of the table to retrieve or a \code{tbl_spark} object.}

\item{sc}{A \code{spark_connection}.}
}
\value{
A \code{tbl_df}.
}
\description{
Given the name of a Spark DataFrame or a reference to a Spark DataFrame, pull
it into R from a \code{spark_connection}.
}
