% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_push_data.R
\name{spark_push_data}
\alias{spark_push_data}
\title{Upload Data to Spark}
\usage{
spark_push_data(sc, x, name, overwrite = FALSE, ...)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{x}{A \code{data.frame}.}

\item{name}{\code{character(1)}. The name to give to the data in the Spark
context.}

\item{overwrite}{\code{logical(1)}.}

\item{...}{Additional arguments to be passed on to
\code{\link[sparklyr:spark_read_csv]{sparklyr::spark_read_csv()}}.}
}
\value{
A \code{tbl_spark} which is a pointer to a Spark DataFrame within \code{sc}
}
\description{
Given a \code{data.frame} in R, upload it to a Spark context.
}
\details{
This function is a faster version of \code{\link[dplyr:copy_to]{dplyr::copy_to()}}.
}
