% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparklio-package.R
\docType{package}
\name{sparklio}
\alias{sparklio}
\alias{sparklio-package}
\title{sparklio: Utility Functions to Extend 'sparklyr'\'s IO}
\description{
A common API provided for working with data between Spark and R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nathaneastwood/sparklio}
  \item Report bugs at \url{https://github.com/nathaneastwood/sparklio/issues}
}

}
\author{
\strong{Maintainer}: Nathan Eastwood \email{nathan.eastwood@icloud.com}

}
\keyword{internal}
