% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_drop.R
\name{spark_drop}
\alias{spark_drop}
\alias{spark_drop_table}
\alias{spark_drop_all_tables}
\title{Drop tables from a Spark session}
\usage{
spark_drop_table(sc, tbl)

spark_drop_all_tables(sc)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{tbl}{\code{character(1)}. The name of the table to remove.}
}
\value{
\code{logical(1)} referring to whether the removal was successful (\code{TRUE})
or not (\code{FALSE}).
}
\description{
This functionality is similar to \code{dbplyr::drop_db_table()} but invokes the
Spark method without the long call stack.
}
